lib/compile: use produceRules in emitNewStructure
Previously used to push directly to structure, which lead to
breadth-first expansion.  Now we effect a depth-first expansion, which
leads to sub-rules being grouped directly under their parent rules.

Another way of looking at this was that we used to have trampolined
recursion through the outer structure loop in Compile.  This change
makes the recusion explicit and immediate, rather than eventual.